+++ lib/src/models/appmetrica_config.dart
@@ -6,7 +6,7 @@ import 'preload_info.dart';
 /// The class contains the starting configuration of the library.
 /// Configuration parameters are applied from the moment the library is initialized. You can set:
 /// * [apiKey] — application API key;
-/// * [advIdentifiersTracking] - (Android only) indicates whether AppMetrica should include advertising identifiers withing its reports.
+/// * [advIdentifiersTracking] - indicates whether AppMetrica should include advertising identifiers withing its reports.
 /// * [anrMonitoring] - whether to capture and send reports about ANRs automatically. The default value is false.
 /// * [anrMonitoringTimeout] - the timeout in seconds at which the fact of ANR is recorded.
 /// * [appBuildNumber] - build number of application.