+++ lib/src/platform/converters/appmetrica_library_adapter_config_converter.dart
@@ -1,0 +1,8 @@
+import '../../models/appmetrica_library_adapter_config.dart';
+import '../pigeon/appmetrica_api_pigeon.dart';
+
+extension ConfigConverter on AppMetricaLibraryAdapterConfig {
+  AppMetricaLibraryAdapterConfigPigeon toPigeon() => AppMetricaLibraryAdapterConfigPigeon(
+    advIdentifiersTracking: advIdentifiersTracking,
+  );
+}